refactor: reorganize barrels and imports
