+++ src/components/admin/ui/index.js
@@ -1,0 +1,11 @@
+import AdminHeader from '@/components/admin/ui/AdminHeader'
+import AdminTable from '@/components/admin/ui/AdminTable'
+import LayoutAdmin from '@/components/admin/ui/LayoutAdmin'
+import SideMenu from '@/components/admin/ui/SideMenu'
+
+export {
+  AdminHeader,
+  AdminTable,
+  LayoutAdmin,
+  SideMenu
+}
