+++ src/components/ui/index.js
@@ -1,11 +1,15 @@
-import IconUi from '@/components/ui/IconUi'
+import BreadcrumbUi from '@/components/ui/BreadcrumbUi'
 import ConfirmUi from '@/components/ui/ConfirmUi'
+import IconUi from '@/components/ui/IconUi'
 import MenuUi from '@/components/ui/MenuUi'
 import MessageUi from '@/components/ui/MessageUi'
+import TabsUi from '@/components/ui/TabsUi'
 
 export {
-  IconUi,
+  BreadcrumbUi,
   ConfirmUi,
+  IconUi,
   MenuUi,
-  MessageUi
+  MessageUi,
+  TabsUi
 }
